Reorganized lib folder
Created language directory to hold the language.rb file and all its dependencies.
