+++ lib/eson/language/lexeme_capture.rb
