+++ lib/eson/language/rule_seq.rb
