+++ test/rule_test.rb
@@ -1,7 +1,7 @@
 require 'minitest/autorun'
 require 'minitest/pride'
 require 'pp'
-require_relative '../lib/eson/language'
+require_relative '../lib/eson/language/language.rb'
 require_relative '../lib/eson/token_pass'
 
 describe Eson::Language::Rule do
